refactor: extract proxy streaming handler
