+++ src/routes/shared/proxy-handler.ts
@@ -5,21 +5,20 @@
  * Delegates to:
  *   - account-acquisition.ts  — acquire / release with idempotent guard
  *   - proxy-error-handler.ts  — CodexApiError classification + pool state mutations
- *   - response-processor.ts   — streaming (SSE) response path
+ *   - streaming-handler.ts    — streaming (SSE) response lifecycle
+ *   - non-streaming-handler.ts — collect / retry response lifecycle
  */
 
 import type { StatusCode } from "hono/utils/http-status";
-import { stream } from "hono/streaming";
 import { CodexApiError } from "../../proxy/codex-api.js";
 import { withRetry } from "../../utils/retry.js";
 import { debugDump, debugDumpEnabled } from "../../utils/debug-dump.js";
 import { acquireAccount, releaseAccount } from "./account-acquisition.js";
 import { handleCodexApiError, toErrorStatus } from "./proxy-error-handler.js";
 import { isPreviousResponseNotFoundError, isUnansweredFunctionCallError } from "../../proxy/error-classification.js";
-import { streamResponse } from "./response-processor.js";
+import { handleStreaming } from "./streaming-handler.js";
 import { handleNonStreaming } from "./non-streaming-handler.js";
 import { annotateImageGenOutcome, buildCodexApi, stripCodexErrorPrefix } from "./proxy-handler-utils.js";
-import type { UsageInfo } from "../../translation/codex-event-extractor.js";
 import type {
   FormatAdapter,
   HandleProxyRequestOptions,
@@ -31,7 +30,6 @@ import { getConfig } from "../../config.js";
 import { jitterInt } from "../../utils/jitter.js";
 import { getSessionAffinityMap } from "../../auth/session-affinity.js";
 import { enqueueLogEntry } from "../../logs/entry.js";
-import { recordStreamCloseEvent } from "../../logs/stream-close-event.js";
 import { randomUUID } from "crypto";
 import { computeVariantHash } from "./variant-hash.js";
 import { getWsPool } from "../../proxy/ws-pool.js";
@@ -153,9 +151,6 @@ export async function handleProxyRequest(options: HandleProxyRequestOptions): Pr
   const triedEntryIds: string[] = [entryId];
   let modelRetried = false;
   let stripAndRetryDone = false;
-  let usageInfo: UsageInfo | undefined;
-  let capturedResponseId: string | null = null;
-  const responseFunctionCallIds = new Set<string>();
   let activeUsageHint: UsageHint | undefined;
   let implicitResumeActive = false;
   // Idempotent-release guard: prevents double-release across retry branches
@@ -345,102 +340,22 @@ export async function handleProxyRequest(options: HandleProxyRequestOptions): Pr
 
       // ── Streaming path ──
       if (req.isStreaming) {
-        c.header("Content-Type", "text/event-stream");
-        c.header("Cache-Control", "no-cache");
-        c.header("Connection", "keep-alive");
-
-        const capturedEntryId = entryId;
-        const capturedApi = codexApi;
-
-        return stream(c, async (s) => {
-          s.onAbort(() => {
-            console.warn(`[stream-client-abort] rid=${requestId.slice(0, 8)} tag=${fmt.tag} model=${req.model}`);
-            recordStreamCloseEvent({
-              kind: "client-abort",
-              requestId,
-              tag: fmt.tag,
-              model: req.model,
-              accountEntryId: capturedEntryId,
-              variantHash,
-              responseId: capturedResponseId ?? null,
-            });
-            abortController.abort();
-          });
-          const recordStreamAffinity = (): void => {
-            if (!capturedResponseId) return;
-            affinityMap.record(
-              capturedResponseId,
-              capturedEntryId,
-              chainConversationId,
-              upstreamTurnState,
-              req.codexRequest.instructions ?? undefined,
-              usageInfo?.input_tokens,
-              Array.from(responseFunctionCallIds),
-              variantHash,
-            );
-          };
-          try {
-            await streamResponse({
-              writer: s,
-              api: capturedApi,
-              response: rawResponse,
-              model: req.model,
-              adapter: fmt,
-              onUsage: (u) => {
-                usageInfo = u;
-                recordStreamAffinity();
-              },
-              tupleSchema: req.tupleSchema,
-              onResponseId: (id) => {
-                capturedResponseId = id;
-                recordStreamAffinity();
-              },
-              usageHint: activeUsageHint,
-              onResponseMetadata: (metadata) => {
-                for (const callId of metadata.functionCallIds ?? []) {
-                  responseFunctionCallIds.add(callId);
-                }
-                recordStreamAffinity();
-              },
-              diagnostics: {
-                requestId: requestId.slice(0, 8),
-                tag: fmt.tag,
-                provider: "codex",
-                path: "/codex/responses",
-                accountEntryId: capturedEntryId,
-                variantHash,
-                abortSignal: abortController.signal,
-              },
-            });
-          } finally {
-            abortController.abort();
-            recordStreamAffinity();
-            if (usageInfo) {
-              const uncached = usageInfo.cached_tokens
-                ? usageInfo.input_tokens - usageInfo.cached_tokens
-                : usageInfo.input_tokens;
-              const imgIn = usageInfo.image_input_tokens ?? 0;
-              const imgOut = usageInfo.image_output_tokens ?? 0;
-              const hitPct = usageInfo.input_tokens > 0
-                ? `${((usageInfo.cached_tokens ?? 0) / usageInfo.input_tokens * 100).toFixed(1)}%`
-                : "n/a";
-              console.log(
-                `[${fmt.tag}] Account ${capturedEntryId} | rid=${requestId.slice(0, 8)} | Usage: in=${usageInfo.input_tokens}` +
-                (usageInfo.cached_tokens ? ` (cached=${usageInfo.cached_tokens} uncached=${uncached})` : "") +
-                ` out=${usageInfo.output_tokens}` +
-                (usageInfo.reasoning_tokens ? ` reasoning=${usageInfo.reasoning_tokens}` : "") +
-                (imgIn || imgOut ? ` image=${imgIn}/${imgOut}` : "") +
-                ` | hit=${hitPct}`,
-              );
-              if (usageInfo.input_tokens > 10_000) {
-                console.warn(
-                  `[${fmt.tag}] ⚠ High input token count: ${usageInfo.input_tokens} tokens` +
-                  (usageInfo.reasoning_tokens ? ` (reasoning=${usageInfo.reasoning_tokens})` : ""),
-                );
-              }
-            }
-            releaseAccount(accountPool, capturedEntryId, annotateImageGenOutcome(usageInfo, req.expectsImageGen), released);
-          }
+        return handleStreaming({
+          c,
+          accountPool,
+          req,
+          fmt,
+          api: codexApi,
+          response: rawResponse,
+          entryId,
+          abortController,
+          released,
+          requestId,
+          affinityMap,
+          conversationId: chainConversationId,
+          turnState: upstreamTurnState,
+          usageHint: activeUsageHint,
+          variantHash,
         });
       }
